+++ arrow-data/src/main/kotlin/arrow/data/utils.kt
@@ -18,5 +18,5 @@ interface GetterOperation<in K, out V> {
 
 class GetterOperationImpl<in K, out V>(override val getter: (K) -> V) : GetterOperation<K, V>
 
-const val DeprecatedUnsafeAccess: String = "This function is unsafe and will be removed in future versions of Arrow"
+const val DeprecatedUnsafeAccess: String = "This function is unsafe and will be removed in future versions of Arrow. Replace or import `arrow.syntax.unsafe.*` if you wish to continue using it in this way"
 const val DeprecatedAmbiguity: String = "This function is ambiguous and will be removed in future versions of Arrow"
