--- funktionale-try/gradle.properties
@@ -1,4 +1,0 @@
-# Maven publishing configuration
-POM_NAME=Try Module
-POM_ARTIFACT_ID=funktionale-try
-POM_PACKAGING=jar
